move motor state to motor_utils
